Use ANDROID_DETECTED
