+++ kotlinx-coroutines-core/js/src/internal/CoroutineExceptionHandlerImpl.kt
@@ -1,10 +1,12 @@
 /*
- * Copyright 2016-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2016-2022 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.coroutines
+package kotlinx.coroutines.internal
 
 import kotlin.coroutines.*
+
+internal actual val ANDROID_DETECTED = false
 
 internal actual fun propagateExceptionToPlatform(context: CoroutineContext, exception: Throwable) {
     // log exception
