+++ kotlinx-coroutines-core/jvm/src/internal/CoroutineExceptionHandlerImplJvm.kt
@@ -1,11 +1,12 @@
 /*
- * Copyright 2016-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2016-2022 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.coroutines
+package kotlinx.coroutines.internal
 
 import java.util.*
 import kotlin.coroutines.*
+import kotlinx.coroutines.*
 
 /**
  * A list of globally installed [CoroutineExceptionHandler] instances.
